Refactor: create a RepositoryFactory interface in domain
- Move current one to data as DefaultRepositoryFactory
