+++ src/data/metadata/__tests__/integration/local-instance-mapped.spec.ts
@@ -7,6 +7,7 @@ import { Instance } from "../../../../domain/instance/entities/Instance";
 import { SynchronizationBuilder } from "../../../../domain/synchronization/entities/SynchronizationBuilder";
 import { startDhis } from "../../../../utils/dhisServer";
 import { AggregatedD2ApiRepository } from "../../../aggregated/AggregatedD2ApiRepository";
+import { DefaultRepositoryFactory } from "../../../common/factories/DefaultRepositoryFactory";
 import { StorageClientD2Repository } from "../../../config/StorageClientD2Repository";
 import { InstanceD2ApiRepository } from "../../../instance/InstanceD2ApiRepository";
 import { InstanceFileD2Repository } from "../../../instance/InstanceFileD2Repository";
@@ -219,7 +220,7 @@ describe("Sync local instance mapped", () => {
 });
 
 function buildRepositoryFactory() {
-    const repositoryFactory: RepositoryFactory = new RepositoryFactory("");
+    const repositoryFactory: RepositoryFactory = new DefaultRepositoryFactory("");
     repositoryFactory.bind(Repositories.InstanceRepository, InstanceD2ApiRepository);
     repositoryFactory.bind(Repositories.ConfigRepository, StorageClientD2Repository);
     repositoryFactory.bind(Repositories.MetadataRepository, MetadataD2ApiRepository);
