+++ src/data/metadata/__tests__/integration/sync-metadata.spec.ts
@@ -7,6 +7,7 @@ import { MetadataPayloadBuilder } from "../../../../domain/metadata/builders/Met
 import { MetadataSyncUseCase } from "../../../../domain/metadata/usecases/MetadataSyncUseCase";
 import { SynchronizationBuilder } from "../../../../domain/synchronization/entities/SynchronizationBuilder";
 import { startDhis } from "../../../../utils/dhisServer";
+import { DefaultRepositoryFactory } from "../../../common/factories/DefaultRepositoryFactory";
 import { StorageClientD2Repository } from "../../../config/StorageClientD2Repository";
 import { InstanceD2ApiRepository } from "../../../instance/InstanceD2ApiRepository";
 import { TransformationD2ApiRepository } from "../../../transformations/TransformationD2ApiRepository";
@@ -218,7 +219,7 @@ describe("Sync metadata", () => {
 });
 
 function buildRepositoryFactory() {
-    const repositoryFactory: RepositoryFactory = new RepositoryFactory("");
+    const repositoryFactory: RepositoryFactory = new DefaultRepositoryFactory("");
     repositoryFactory.bind(Repositories.InstanceRepository, InstanceD2ApiRepository);
     repositoryFactory.bind(Repositories.ConfigRepository, StorageClientD2Repository);
     repositoryFactory.bind(Repositories.MetadataRepository, MetadataD2ApiRepository);
